Update the sample code to include two cases of StateContainers, a Store and a ViewModel
