+++ app/src/main/res/values/styles.xml
@@ -51,14 +51,11 @@
         <item name="android:textColorHint">@color/grey</item>
     </style>
 
-    <style name="AppTheme.Text.AppButton" parent="AppTheme.Text">
-        <item name="android:textAllCaps">true</item>
+    <style name="AppTheme.AppButton" parent="Widget.AppCompat.Button">
         <item name="android:gravity">center</item>
-        <item name="android:textStyle">bold</item>
-        <item name="android:textSize">@dimen/text_size_medium</item>
-        <item name="android:layout_height">48dp</item>
+        <item name="android:textSize">@dimen/text_size_default</item>
         <item name="android:textColor">@color/colorPrimaryText</item>
-        <item name="android:background">@color/colorAccent</item>
+        <item name="android:backgroundTint">@color/colorAccent</item>
     </style>
 
 </resources>
